[Add] 서비스 컨트롤러 수정 및 추가
issue : #S02P31D107-10, #S02P31D107-11
story : #S02P31D107-49 ~65
- 모든 컨트롤러, 레포지토리, 서비스 수정 및 추가
- Id를 Int > Long으로 변경
- CRUD기능추가

+++ backend/learnacademy/src/main/kotlin/com/ssafy/learnacademy/controller/FreeBoardController.kt
@@ -16,7 +16,7 @@ class FreeBoardController(var freeBoardService:FreeBoardService){
 
     @GetMapping("/{freeBoardId}")
     @ApiOperation(value = "자유게시판 검색", notes = "자유게시판을 검색합니다")
-    fun getFreeBoard(@PathVariable("freeBoardId") freeBoardId : Int) : Optional<FreeBoard>?{
+    fun getFreeBoard(@PathVariable("freeBoardId") freeBoardId : Long) : Optional<FreeBoard>?{
         return freeBoardService.findById(freeBoardId)
     }
 
@@ -34,7 +34,7 @@ class FreeBoardController(var freeBoardService:FreeBoardService){
 
     @DeleteMapping("/{freeBoardId}")
     @ApiOperation(value = "자유게시판 삭제",notes = "자유게시판을 삭제합니다")
-    fun deleteFreeBoard(@PathVariable("freeBoardId") freeBoardId : Int){
+    fun deleteFreeBoard(@PathVariable("freeBoardId") freeBoardId : Long){
         return freeBoardService.deleteFreeBoard(freeBoardId)
     }
 }
